Create LICENSE
Added an MIT License to the project by creating a `LICENSE` file in the root directory. This license grants users the freedom to use, modify, and distribute the software, provided they include the original copyright notice and disclaimers. Including this license clarifies the terms under which the project can be utilized and encourages open-source collaboration.